Move sentinels to their own files
